solve flake8 issues

+++ citam/engine/floorplan_utils.py
@@ -150,8 +150,8 @@ def get_aisle_center_point_and_width(aisle):
     V_perp = gsu.calculate_perpendicular_vector(aisle[0], aisle[1])
     # Half way between the two walls from the middle of the first wall
     mid_point = aisle[0].point(0.5)
-    center_point = Point(x=round(mid_point.real + V_perp[0]/2.0),
-                         y=round(mid_point.imag + V_perp[1]/2.0)
+    center_point = Point(x=int(round(mid_point.real + V_perp[0]/2.0)),
+                         y=int(round(mid_point.imag + V_perp[1]/2.0))
                          )
     # Width of aisle
     width = (V_perp[0]**2 + V_perp[1]**2)**0.5
